added PIGGG

+++ Sources/Resources/PlayerTexture.cpp
@@ -12,7 +12,7 @@ void Resources::PlayerTexture::Load(const char* playerName)
 {
 	fullPath = playerName;
 	if (loaded.Load()) return;
-	int max = fullPath.substr(15).copy(Name, 255);
+	int max = (int)fullPath.substr(15).copy(Name, 255);
 	Name[max] = 0;
 
 	// TODO: fill data
